# docs/admin/swagger.yaml
basePath: /api/admin
definitions:
  base_info.Success:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      msg:
        example: err msg
        type: string
      reason:
        description: 错误原因，定义为业务判定错误码
        example: success
        type: string
    type: object
  base_info.Swagger400Resp:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger401Resp:
    properties:
      code:
        example: 401
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger500Resp:
    properties:
      code:
        example: 500
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
...
  commands.AssignUsersToDepartmentCommand:
    properties:
      deptId:
        type: string
      userIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - deptId
    - userIds
    type: object
...
  commands.CreateUserCommand:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      name:
        description: 名称
        type: string
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        minLength: 6
        type: string
      phone:
        description: 手机号
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
...
  commands.RemoveUsersFromDepartmentCommand:
    properties:
      deptId:
        type: string
      userIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - deptId
    - userIds
    type: object
...
  commands.UpdateUserCommand:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        description: 名称
        type: string
      nickname:
        type: string
      phone:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    type: object
  commands.UpdateUserStatusCommand:
    properties:
      id:
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    - status
    type: object
...
  dto.PermissionsTreeDto:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.PermissionsTreeDto'
        type: array
      code:
        description: 编码
        type: string
      description:
        description: 描述
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: integer
      localize:
        description: 本地化
        type: string
      name:
        description: 名称
        type: string
      parent_path:
        description: 父级路径
        type: string
      parentId:
        description: 父级ID
        type: integer
      path:
        description: 路径
        type: string
      properties:
        description: 属性
        type: string
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      type:
        description: 类型
        type: integer
    type: object
...
  dto.UserDto:
    properties:
      avatar:
        description: 头像
        type: string
      createdAt:
        description: 创建时间
        type: integer
      email:
        description: 邮箱
        type: string
      id:
        description: 用户ID
        type: string
      invitationCode:
        description: 邀请码
        type: string
      name:
        description: 姓名
        type: string
      nickname:
        description: 昵称
        type: string
      phone:
        description: 手机号
        type: string
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      status:
        description: 状态,1启用,2禁用
        type: integer
      tenantId:
        description: 租户ID
        type: string
      updatedAt:
        description: 更新时间
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  dto.UserInfoDto:
    properties:
      homePage:
        description: 首页
        type: string
      permissions:
        description: 所有权限列表
        items:
          type: string
        type: array
      roles:
        description: 角色列表
        items:
          type: string
        type: array
      user:
        $ref: '#/definitions/dto.UserDto'
    type: object
...
  models.PageRes-dto_UserDto:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.UserDto'
        type: array
      total:
        type: integer
    type: object
host: localhost:8888
info:
  contact:
    name: go-server-template-admin
  description: This is a demo using go-server-template-admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ares-ddd-admin
  version: "1.0"
paths:
...
  /v1/sys/dept/{id}/users:
    get:
      consumes:
      - application/json
      description: 获取部门下的用户列表(分页)
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 姓名
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/models.PageRes-dto_UserDto'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取部门用户列表
      tags:
      - 部门管理
...
  /v1/sys/dept/users:
    delete:
      consumes:
      - application/json
      description: 从部门移除用户
      parameters:
      - description: 移除用户参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.RemoveUsersFromDepartmentCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 从部门移除用户
      tags:
      - 部门管理
    post:
      consumes:
      - application/json
      description: 分配用户到部门
      parameters:
      - description: 分配用户参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.AssignUsersToDepartmentCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 分配用户到部门
      tags:
      - 部门管理
...
  /v1/sys/user:
    get:
      description: 获取用户
      operationId: UserList
      parameters:
      - description: 页码，从1开始
        in: query
        name: current
        type: integer
      - in: query
        name: email
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: phone
        type: string
      - description: 页码大小，最大100
        in: query
        name: size
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: username
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取用户
      tags:
      - 系统用户
    post:
      description: 添加用户
      operationId: AddUser
      parameters:
      - description: 属性说明请在对应model中查看
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.CreateUserCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 添加用户
      tags:
      - 系统用户
    put:
      consumes:
      - application/json
      description: 更���用户信息，包括基本信息和用户关联
      operationId: UpdateUser
      parameters:
      - description: 用户更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateUserCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 更新用户
      tags:
      - 系统用户
  /v1/sys/user/:id:
    get:
      consumes:
      - application/json
      description: 根据id获取用户
      operationId: GetUserDetails
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDto'
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 根据id获取用户
      tags:
      - 系统用户
  /v1/sys/user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的用户
      operationId: DeleteUser
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 删除用户
      tags:
      - 系统用户
  /v1/sys/user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息，包括权限和菜单
      operationId: GetUserInfo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfoDto'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取用户信息
      tags:
      - 系统用户
  /v1/sys/user/menus:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的菜单树
      operationId: GetUserMenus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsTreeDto'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取用户菜单树
      tags:
      - 系统用户
  /v1/sys/user/status:
    put:
      consumes:
      - application/json
      description: 更新用户的启用/禁用状态
      operationId: UpdateUserStatus
      parameters:
      - description: 用户状态更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.UpdateUserStatusCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 更新用户状态
      tags:
      - 系统用户
schemes:
- http
swagger: "2.0"